services/auth-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
passlib
python-jose[cryptography]
python-multipart